# contracts/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "VNI IELTS AI",
    "description": "The exam engine, identity, and content APIs. Generated from the running application \u2014 never hand-edited.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/library/documents": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "Published documents; filter by skill, type, band, or free text",
        "operationId": "ListLibraryDocuments",
        "parameters": [
          {
            "name": "skill",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "band",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDocumentListView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        }
      }
    },
    "/api/v1/library/documents/{id}": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "One published document",
        "operationId": "GetLibraryDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDocumentView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        }
      }
    },
    "/api/v1/library/articles": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "Published articles, newest first; filter by category",
        "operationId": "ListArticles",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleListView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        }
      }
    },
    "/api/v1/library/articles/{slug}": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "One published article by its slug; a draft slug is 404",
        "operationId": "GetArticleBySlug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        }
      }
    },
...
    "/api/v1/admin/library/documents": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Every document, drafts included",
        "operationId": "AdminListLibraryDocuments",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDocumentListView"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create a draft document",
        "operationId": "AdminCreateLibraryDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryDocumentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDocumentView"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/library/documents/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "AdminGetLibraryDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDocumentView"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Replace a document\u0027s editable fields; status is unchanged",
        "operationId": "AdminUpdateLibraryDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryDocumentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDocumentView"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete a document that is not published",
        "operationId": "AdminDeleteLibraryDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/library/articles": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Every article, drafts included",
        "operationId": "AdminListArticles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleListView"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create a draft article; 409 SLUG_TAKEN if the slug is in use",
        "operationId": "AdminCreateArticle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleView"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/library/articles/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "AdminGetArticle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleView"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Replace an article\u0027s editable fields; status is unchanged",
        "operationId": "AdminUpdateArticle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleView"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete an article that is not published",
        "operationId": "AdminDeleteArticle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "No token, or one this API will not accept."
          },
          "429": {
            "description": "Rate limited. Always carries Retry-After \u2014 a 429 without it invites an immediate retry, which makes the problem worse."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArticleInput": {
        "required": [
          "slug",
          "title",
          "excerpt",
          "category",
          "readMinutes",
          "author",
          "body"
        ],
        "type": "object",
        "properties": {
          "slug": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "excerpt": {
            "type": [
              "null",
              "string"
            ]
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "readMinutes": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "author": {
            "type": [
              "null",
              "string"
            ]
          },
          "body": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ArticleListView": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleSummaryView"
            }
          }
        }
      },
      "ArticleSummaryView": {
        "required": [
          "id",
          "slug",
          "title",
          "excerpt",
          "category",
          "readMinutes",
          "author",
          "relatedExamIds",
          "status",
          "createdAt",
          "updatedAt",
          "publishedAt",
          "createdBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "readMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "author": {
            "type": "string"
          },
          "relatedExamIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publishedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          }
        }
      },
      "ArticleView": {
        "required": [
          "id",
          "slug",
          "title",
          "excerpt",
          "category",
          "readMinutes",
          "author",
          "body",
          "relatedExamIds",
          "status",
          "createdAt",
          "updatedAt",
          "publishedAt",
          "createdBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "readMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "author": {
            "type": "string"
          },
          "body": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "relatedExamIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publishedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          }
        }
      },
...
      "LibraryDocumentInput": {
        "required": [
          "slug",
          "title",
          "description",
          "skill",
          "category",
          "type",
          "format",
          "targetBand",
          "topic",
          "pageCount",
          "size",
          "fileUrl",
          "isFeatured",
          "isNew",
          "isUpdated",
          "isPopular",
          "access"
        ],
        "type": "object",
        "properties": {
          "slug": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "skill": {
            "type": [
              "null",
              "string"
            ]
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "format": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetBand": {
            "type": [
              "null",
              "string"
            ]
          },
          "topic": {
            "type": [
              "null",
              "string"
            ]
          },
          "pageCount": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "size": {
            "type": [
              "null",
              "string"
            ]
          },
          "fileUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "isFeatured": {
            "type": "boolean"
          },
          "isNew": {
            "type": "boolean"
          },
          "isUpdated": {
            "type": "boolean"
          },
          "isPopular": {
            "type": "boolean"
          },
          "access": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "LibraryDocumentListView": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryDocumentView"
            }
          }
        }
      },
      "LibraryDocumentView": {
        "required": [
          "id",
          "slug",
          "title",
          "description",
          "skill",
          "category",
          "type",
          "format",
          "targetBand",
          "topic",
          "pageCount",
          "size",
          "fileUrl",
          "isFeatured",
          "isNew",
          "isUpdated",
          "isPopular",
          "access",
          "relatedExamIds",
          "status",
          "createdAt",
          "updatedAt",
          "publishedAt",
          "createdBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "skill": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "targetBand": {
            "type": [
              "null",
              "string"
            ]
          },
          "topic": {
            "type": [
              "null",
              "string"
            ]
          },
          "pageCount": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "size": {
            "type": "string"
          },
          "fileUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "isFeatured": {
            "type": "boolean"
          },
          "isNew": {
            "type": "boolean"
          },
          "isUpdated": {
            "type": "boolean"
          },
          "isPopular": {
            "type": "boolean"
          },
          "access": {
            "type": "string"
          },
          "relatedExamIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publishedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "The access token from /api/v1/auth/login, /auth/refresh or the SSO handoff. Fifteen minutes; the client refreshes ahead of expiry.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}